Added adujstement to Applications.scss, dash board buttons, acceessory form warning debug
