+++ app/assets/stylesheets/application.scss
@@ -12,8 +12,9 @@
 @import "pages/index";
 
 .main-yield-container {
-  margin-top: 3em;
-  padding-left: 1em;
-  padding-right: 1em;
   // border: 1px solid black;
+  margin-top: 4rem;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  margin-bottom: 5rem;
 }
