+++ app/javascript/application.js
@@ -3,4 +3,3 @@ import "@hotwired/turbo-rails"
 import "controllers"
 import "@popperjs/core"
 import "bootstrap"
-import "accessory_form"
