+++ app/views/accessories/new.html.erb
@@ -1,5 +1,11 @@
+<%= link_to 'Back', accessories_path, class: 'btn btn-secondary' %>
 <h1>New Accessory</h1>
 
 <%= render 'form', accessory: @accessory %>
 
-<%= link_to 'Back', accessories_path, class: 'btn btn-secondary' %>
+<script type="module">
+  import { initAccessoryForm } from 'controllers/accessory_form';
+  document.addEventListener('DOMContentLoaded', () => {
+    initAccessoryForm();
+  });
+</script>
