+++ app/views/pages/dashboard.html.erb
@@ -6,8 +6,8 @@
     <div class="card-header d-flex justify-content-between align-items-center">
       <h2>Recent Extractions</h2>
       <div>
-        <%= link_to 'New Extraction', new_extraction_path, class: 'btn btn-primary mx-1' %>
-        <%= link_to 'View All Extractions', extractions_path, class: 'btn btn-secondary mx-1' %>
+        <%= link_to 'New', new_extraction_path, class: 'btn btn-primary mx-1' %>
+        <%= link_to 'View All', extractions_path, class: 'btn btn-secondary mx-1' %>
       </div>
     </div>
     <div class="card-body">
@@ -32,8 +32,8 @@
     <div class="card-header d-flex justify-content-between align-items-center">
       <h2>Recent Coffees</h2>
       <div>
-        <%= link_to 'New Coffee', new_user_coffee_path, class: 'btn btn-primary mx-1' %>
-        <%= link_to 'View All Coffees', user_coffees_path, class: 'btn btn-secondary mx-1' %>
+        <%= link_to 'New', new_user_coffee_path, class: 'btn btn-primary mx-1' %>
+        <%= link_to 'View All', user_coffees_path, class: 'btn btn-secondary mx-1' %>
       </div>
     </div>
     <div class="card-body">
@@ -58,8 +58,8 @@
   <div class="card-header d-flex justify-content-between align-items-center">
     <h2>Accessories</h2>
     <div>
-      <%= link_to 'New Accessory', new_accessory_path, class: 'btn btn-primary mx-1' %>
-      <%= link_to 'View All Accessories', accessories_path, class: 'btn btn-secondary mx-1' %>
+      <%= link_to 'New', new_accessory_path, class: 'btn btn-primary mx-1' %>
+      <%= link_to 'View All', accessories_path, class: 'btn btn-secondary mx-1' %>
     </div>
   </div>
   <div class="card-body">
