+++ config/importmap.rb
@@ -7,4 +7,5 @@ pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
 pin "bootstrap", to: "bootstrap.min.js", preload: true
 pin "@popperjs/core", to: "popper.js", preload: true
-pin "accessory_form", to: "accessory_form.js"
+pin "controllers/accessory_form", to: "controllers/accessory_form.js"
+
